journalserver/requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-admin==2.0.2
django-allauth==65.13.0
django-cors-headers==4.9.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
excel-base==1.0.4
google-auth==2.42.1
idna==3.11
isoweek==1.3.3
jwt==1.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
xlwt==1.3.0
